move the last of the net tests
